extract parseImageDimensions to ignore typecheck for the entire function
